Fix errors in tacker compliance tests
This patch fixes the bug in tacker compliance test regarding
some tests cases get a failure saying "None is not of type
'string'".

As a solution we have implemented - "After reading value from
tacker DB, if "vnfInstanceDescription" has the value "None",
then omit "vnfInstanceDescription" key and its value. We have
implemented this solution because it meets the ETSI specification.

Other solution is modify json schema file in NFV-TST code so that
"null" type is also acceptable like below.

  "vnfInstanceDescription": {
     "description": "Human-readable description of the VNF instance.",
     "type": ["string", "null"]
   }

-> Add "null" type into "vnfInstanceDescription schema definition.

But, according to NFV-SOL003 specification, the data type of
"vnfInstanceDescription" is defined only as "string". This means
adding "null" type is out of ETSI specification.

Therefore, we have implemented fix in Tacker's side because fix can
not be handled at NFV-TST side.

Closes-Bug: #2021454
Change-Id: Ie058285d15c0542624977f9fdf5a8efaf53c8373