Fix the error caused by heal all failed
When heal all(all=true) is executed, the stack id will be updated.
Once the stack id is not saved in vnf_instance, it will cause an
exception.
In this patch, in heal and terminate, if the stack id does not
exist, use the stack name instead to call the heat api.

Closes-Bug: #2032883
Change-Id: I5d205432ffb6574678b7e1846819e2683a36d89b