Set system scope in oslo policy creds object
In tacker.policy.authorize() method, it convert the context
object to policy value by context.to_policy_values() and then
pass that into oslo.policy enforcer authorize() method.

This is fine till now and does not cause any issue but as per
SRABC, we need to set if token is system scope (for project
scope it is all fine) in oslo policy creds via key 'system'.
But context.to_policy_values() method does not set the 'system'
key in creds because there it is named as `system_scope`.

To fix that we need to explicitly set the 'system' key in creds.

Partial implement blueprint implement-project-personas

Change-Id: I65d28749a4584661f7f4f596c4d2c39a84730963