Add testing framework for RBAC
Current tests do not have good test coverage of existing
RBAC policies. Either tests for policies do not exist or
if they exist then they do not cover the actual negative
and positive testing.

Along with what all users have access, it is important to
test what all users does not have access.

For Example, if any policy with default rule as admin only then
test should verify:
- policy check pass with context having admin role
- policy check fail with context having non-admin role

As we are implementing the project personas (project member
and reader role) in policies, we need to have the enough
testing coverage of existing policy behavior and to know
that with new defaults how the access permissions will
looks like. These test coverage will be extended to adopt
the new changes and also make sure we do not break the
existing behavior.

This commit adds the testing framework for RBAC and implement
create VNF test as example. It will cover other APIs in further
changes in this series.

Partial implement blueprint implement-project-personas

Change-Id: I5b0d039c6aebda6ba0653032ac5a1963a704cb59