Fix output regarding PM Job/Threshold response
There are two corrections:

First, according to SOL003, "authentication" should not appear
in the PmJobModifications response.
Therefore, it has been removed from the output obtained
when running "openstack vnfpm job update".

Second, the data type of PmJob href, as defined in SOL003,
has been changed to match the format of the Uri and
Link href defined in SOL013. This adjustment ensures
that the endpoints will now be displayed correctly.

Closes-Bug: #2054420
Change-Id: Id79ed19090271de4e32742f37cf54b6853acba2e