Performance enhancement of v2 list APIs
This patch enhances performance of the following APIs.
* GET /vnflcm/v2/vnf_instances
* GET /vnflcm/v2/vnf_lcm_op_occs

Previously a DB object was converted into a tacker object and
then it was converted into dict, but each conversion took quite
long time. This patch fixes to convert a DB object into dict
directly.

In addition, the query for DB is improved to apply filter and
selector specified in the query patameter of the API when its
condition is simple.

There are some refactors with improvement of viewer too.

Implements: blueprint perf-improve-list-apis
Change-Id: I5e821c6363290623356df38c3bd43054b54f89fd