Fix tox debug not work for invalid config register
Some unittests cannot be debugged because of failures occurred while
running oslo_debug_helper launched from `tox -e debug`. The cause is
a wrong registering configs and it's ended up with
`oslo_config.cfg.DuplicatedOptError`. If you attempt to set an option
twice with the same key but different values, it fails.

This failure only happens while running debug and all the test passes
if you just run the tests for now because the order of importing
modules is different among test scenarios. So, if someone add or
change registering options without taking care about, the failure
will be happened again.

In addition, oslo_debug_helper launched from `tox -e debug` has also
a problem for some deprecated warning which causes an unexpected
failure and failed to run debugger as a result.

This update is a temporary fix and some more update is required as
noted below. It just:
* Removes useless registering options.
* Corrects the values of duplicated options to be the same.
* Adds a option for oslo_debug_helper to avoid to be failed to
  launch.

NOTE: The actual cause is registering config is not organized in
tacker's code. We should revise the way of registering for
oslo_config later.

Closes-Bug: #2077355
Change-Id: Idfc562804a273961ed94e7ae3cf313e3ccaab2b1
Signed-off-by: Yasufumi Ogawa <yasufum.o@gmail.com>