Fix Kubernetes service resource check
This patch fixes the Kubernetes service resource check performed
after instantiate of CNF. It modifies the resource check to retry
if the endpoint is not created after the service has been created
because it may take time a bit to create the endpoint. This error
may occur when the VNF package to be instantiated is large.

Also, this patch fixes some typos in the V1 API, changing
"endpoinds" to "endpoints".

Closes-Bug: #2078288
Change-Id: I3f1b95c72a1ca0943ca13b8b4b56ace181e53598