Generate request files with actual VIM's params
VNF package generator doesn't take care the params of
`vimConnectionInfo`, `ssl_ca_cert` and `bearer_token`, in the
following request files while generating the request files and fill
the fields with dummy params. So, you need to update them by
yourself before using the files. This update is to generate the
request files with available params.

* `instantiate_req` for ETSINFV.OPENSTACK_KEYSTONE.V_3
* `max_sample_instantiate` for ETSINFV.KUBERNETES.V_1
* `helm_instantiate_req` for ETSINFV.HELM.V_3

To give the params to the package generator, this patch introduces
three options, using VIM ID, VIM name or config file.

  $ python3 tools/gen_vnf_pkg.py -t VIM_TYPE -c VIM_CONF

However, it's not supported all the options for every types currently.
For 'ETSINFV.KUBERNETES.V_1' and 'ETSINFV.HELM.V_3', VIM ID and VIM
name are not available. If you force to use, default dummy params are
used instead. The detail info is described in help message.

Change-Id: I91a06dad3d02aaf612d612c66c5341212a029b8f