Add exclusive processing to get_vnfd in v2 API
When using an external NFVO, if VNF instances with the same vnfd_id
are instantiated at the same time, one of the VNF instances will
fail to create the CSAR cache directory and will have a "ROLLED_BACK"
status.
This patch fixes the get_vnfd method so that it cannot be executed on
the same vnfd_id at the same time.

Closes-Bug: #2088124
Change-Id: I15dab7796b356376f39c9a46410f7edb9d2d8242