Improve the ft-v2-st-userdata-err-handling job
This patch removes the check of the vnflcm op list response header
in test_rollback_instantiate of the
tacker-ft-v2-st-userdata-err-handling job.
When test_rollback_instantiate is executed first in a job,
there is only one opocc, so the Link header is not included
in the vnflcm op list response header.
As a result, the test may fail with the following error.
(If the order is other than first, no error occurs.)
  * Exception: Supported header doesn't exist: Link

The Link header check is also performed in other tests,
so there is no serious problem in deleting it here.

Change-Id: I83db8a72a5c9e2b28268c8c28227471c3d13292b