Add tutorial to test external oauth2 server
This patch adds script to start Keycloak authorization server as a
docker container for testing external oauth2 server authentication
and a guide of how to use that script in "Using External OAuth 2.0
Authorization Server for Tacker APIs" documentation.

And the content of the documentation is reconfirmed with 2024.2
release of Tacker.

Change-Id: Id10343da43e121c52227f6c18d0b17af88c72c62