Update "Manual Installation" document
This patch updates the content of "Manual Installation" document by
confirming with the 2024.2 Dalmatian release of Tacker.
And the guide is changed from installing with root user to installing
with general user.
And guide to install python-tackerclient is removed because it will be
installed as requirement while installing Tacker.
Guide to install tacker-horizon is also removed as it has its own
documentation and is not essential to install Tacker server.

Change-Id: Ibdaa6001e2c18738b89b4bce38465a12717e6468