Add tacker_cli for external OAuth2.0
This patch adds CLI shell script which is a wrapper of the curl command
for cases where an external OAuth2.0 authorization server is
configured. Without this script, users need to get OAuth2.0 token and
build a curl command with proper headers and a body manually as the
usual openstack command only work for Keystone.

Change-Id: Idaa159d0f5cc9cc245e141eb595bb2f15c571d03