Add documentation for running v1 FT locally
This patch adds documentation for running v1 FT locally on a multi-node
system similar to Zuul. The porpose of this change is to encourage
users to test Tacker v1 features on local since most of v1 features are
not under active development anymore and planned to be moved to
periodic jobs or dropped from zuul tests.

Five test scenarios are supported in this change as a first step.

  * tacker-ft-legacy-vim
  * tacker-ft-v1-vnfpkgm
  * tacker-ft-v1-k8s
  * tacker-ft-v1-tosca-vnflcm
  * tacker-ft-v1-userdata-vnflcm

It also provides sample DevStack configuration files, setup scripts and
`local.conf.HOSTNAME`.

Implements: blueprint reduce-ft-time
Co-Authored-By: Ai Hamaho <ai.hamano@ntt-at.co.jp>
Co-Authored-By: Yasufumi Ogawa <yasufum.o@gmail.com>
Change-Id: Id0debea403ccb9490da14b1a77c89ff7b42ee178